+++ frontend/public/i18n/de.json
@@ -1,5 +1,6 @@
 {
+  "about": "Über",
   "app-title": "AGE Verwaltungstool",
-  "home": "Startseite",
-  "about": "Über"
+  "experimental": "Experimentell",
+  "home": "Start"
 }
