+++ frontend/public/i18n/en.json
@@ -1,5 +1,6 @@
 {
+  "about": "About",
   "app-title": "AGE Administration Tool",
-  "home": "Home",
-  "about": "About"
+  "experimental": "Experimental",
+  "home": "Home"
 }
